Use pathlib so file gets closed on error
If an error is thrown between `open` and `close` on a Python file, the file might not be closed. Using a `with` block, `try`/`finally` or offloading it to a library that will do those things for you is the convention in Python. `pathlib` is a library designed for manipulating files abstractly and safely.